final version (KL)

+++ AR/Assets/Scripts/jsonModel/Rotation.cs
@@ -2,9 +2,9 @@ namespace ModelData{
 
     public class Rotation
     {
-        public int x { get; set; }
-        public int y { get; set; }
-        public int z { get; set; }
+        public float x { get; set; }
+        public float y { get; set; }
+        public float z { get; set; }
     }
 
 }
